requirements.txt
asgiref==3.5.0
astroid==2.9.3
beautifulsoup4==4.10.0
colorama==0.4.4
coverage==6.3.2
Django==4.0.3
django-filter==21.1
django-guardian==2.4.0
django-htmlmin==0.11.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
html5lib==1.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.12.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
webencodings==0.5.1
wrapt==1.13.3
